requirements.txt
browser-use
chromadb
croniter
discord.py
langchain-google-genai
langchain-mcp-adapters
langchain-ollama
langchain-openai
langgraph
playwright
pytest
python-dotenv
